General code tweaking, nothing strange.

+++ lib/checkbufferoverrun.cpp
@@ -1246,7 +1246,7 @@ void CheckBufferOverrun::checkReadlinkBufferUsage(const Token* tok, const Token 
 void CheckBufferOverrun::checkGlobalAndLocalVariable()
 {
     // check all known fixed size arrays first by just looking them up
-    for (size_t i = 1; i <= _tokenizer->varIdCount(); i++) {
+    for (unsigned int i = 1; i <= _tokenizer->varIdCount(); i++) {
         const Variable *var = _tokenizer->getSymbolDatabase()->getVariableFromVarId(i);
         if (var && var->isArray() && var->dimension(0) > 0) {
             ArrayInfo arrayInfo(var, _tokenizer);
@@ -2100,7 +2100,7 @@ void CheckBufferOverrun::executionPaths()
 {
     // Parse all variables and extract array info..
     std::map<unsigned int, ArrayInfo> arrayInfo;
-    for (size_t i = 1; i <= _tokenizer->varIdCount(); i++) {
+    for (unsigned int i = 1; i <= _tokenizer->varIdCount(); i++) {
         const Variable *var = _tokenizer->getSymbolDatabase()->getVariableFromVarId(i);
         if (var && var->isArray() && var->dimension(0) > 0)
             arrayInfo[i] = ArrayInfo(var, _tokenizer);
